% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{CalculateThreshold}
\alias{CalculateThreshold}
\title{Backward-compatible wrapper for \code{calculate_threshold()}}
\usage{
CalculateThreshold(data, meteo, dates = NULL, var, percent = 95, plot = FALSE)
}
\arguments{
\item{data}{A standard bat contact data frame.}

\item{meteo}{A weather data frame.}

\item{dates}{Optional vector of start and end dates.}

\item{var}{Character vector of weather variables.}

\item{percent}{Percentile to calculate for each variable.}

\item{plot}{Logical. Whether to print diagnostic plots.}
}
\value{
A data frame of threshold model results.
}
\description{
Backward-compatible wrapper for \code{calculate_threshold()}
}
