% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microphone.R
\name{plot_microphone_test}
\alias{plot_microphone_test}
\title{Plot a TeensyRecorder microphone test}
\usage{
plot_microphone_test(test)
}
\arguments{
\item{test}{Object returned by \code{read_microphone_test()}.}
}
\value{
A ggplot object.
}
\description{
Plot a TeensyRecorder microphone test
}
